test: compare the choropleth against committed screenshots
Chromatic archives a serialization of the DOM, which carries a <canvas>
element's size and position and none of its pixels.  The geography
section is drawn by MapLibre into a WebGL canvas, so a map that failed to
draw at all is indistinguishable there from one that drew perfectly -
which is exactly the regression this branch started from.

Playwright takes its screenshots in the browser and does capture canvas
pixels, so a second suite compares the section against PNGs committed
beside it, at a desktop and a phone width in both themes.  Text is masked
out of the comparison, fonts being the least portable thing in the image.

The image diff alone would not be enough: at world zoom the fixture's
three regions come to about a thousand pixels, against a difference
tolerance of some thirty thousand that is what keeps the comparison
stable across machines.  So the map is also asked outright how many
regions it painted.  How many it should be showing depends on the width -
the view is centered on longitude 0, and a phone-width map does not reach
California - so each viewport states its own count.

The page mocks move to tests/fixtures so both suites share them.
