+++ .github/workflows/test.yml
@@ -36,6 +36,41 @@ jobs:
                   fail_ci_if_error: false
                   verbose: true
 
+    visual-tests:
+        name: Visual tests (Playwright screenshots)
+        runs-on: ubuntu-latest
+        permissions:
+            contents: read
+        steps:
+            - name: Checkout
+              uses: actions/checkout@v7
+
+            - name: Set up Node.js
+              uses: actions/setup-node@v7
+              with:
+                  node-version: lts/*
+                  cache: npm
+
+            - name: Install dependencies
+              run: npm ci
+
+            - name: Install Playwright browsers
+              run: npx playwright install chromium --with-deps
+
+            - name: Run visual tests
+              run: npm run test:visual
+
+            # The report holds the expected, actual and diff images of any
+            # comparison that failed, which is what a baseline is regenerated
+            # from when a change to the map is meant rather than a regression.
+            - name: Upload report
+              if: failure()
+              uses: actions/upload-artifact@v4
+              with:
+                  name: playwright-report-visual
+                  path: playwright-report-visual/
+                  retention-days: 7
+
     integration-tests:
         name: Integration tests (Playwright)
         runs-on: ubuntu-latest
