+++ package.json
@@ -16,6 +16,7 @@
     "test:coverage": "vitest run --coverage --config configs/vitest.config.js",
     "test:e2e": "playwright test --config configs/playwright.config.js",
     "test:chromatic": "playwright test --config configs/playwright.chromatic.config.js",
+    "test:visual": "playwright test --config configs/playwright.visual.config.js",
     "test:e2e:coverage": "PW_COVERAGE=1 playwright test --config configs/playwright.config.js",
     "coverage:integration": "node src/scripts/build-integration-coverage.mjs",
     "storybook": "storybook dev -p 6006 --config-dir configs/storybook",
